итог

+++ 1/homework 1.3.py
@@ -1,4 +1,4 @@
-x = int(input("Введите x"))
-y = int(input("Введите y"))
-z = int(input("Введите z"))
-print((x**2+y**2+z**2)**0.5)
+x = float(input("Введите x"))
+y = float(input("Введите y"))
+z = float(input("Введите z"))
+print((x**2 + y**2 + z**2)**0.5)
